Struct tfidf::idf::InverseFrequencyIdf

pub struct InverseFrequencyIdf;

Inverse frequency weighting scheme for IDF. Computes log (N / nt) where N is the number of documents, and nt is the number of times a term appears in the corpus of documents.