Module tfidf::tf

Implementations of different weighting schemes for term frequency (tf). For more information about which ones are implemented, check the Wiki link in the crate description.

Structs

BinaryTf
Binary weighting scheme for TF. If the document contains the term, returns 1, otherwise returns 0.

DoubleHalfNormalizationTf
Double normalized weighting scheme for TF based on a factor, K = 0.5.

LogNormalizationTf
Log normalized weighting scheme for TF. Computes 1 + log (f) where f is the frequency of the term in the document.

RawFrequencyTf
Raw frequency weighting scheme for TF. Returns the number of times a term occurs in the document.

Traits

DoubleKNormalizationTf
Double normalized weighting scheme for TF based on a factor, K.