Trait tfidf::ProcessedDocument

pub trait ProcessedDocument: Document {
    fn term_frequency<K>(&self, term: K) -> usize where K: Borrow<Self::Term>;
    fn max(&self) -> Option<&Self::Term>;
}

A document where the frequencies of each term is already calculated.

Required Methods

fn term_frequency<K>(&self, term: K) -> usize where K: Borrow<Self::Term>

Returns the number of times a (non-normalized) term exists within the document.

fn max(&self) -> Option<&Self::Term>

Returns the term with the highest frequency, or tied for the highest frequency.